SongGenIA-login/backend/Login_Register/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
passlib==1.7.4
pyjwt==2.8.0
pymysql==1.1.0
Flask-Login==0.6.3
Authlib==1.2.1  
bcrypt==4.0.1
flask-cors
Flask-OAuthlib==0.9.5
